добавлена 3 лаба

+++ src/main/java/com/example/labs/service/AuthorService.java
@@ -1,6 +1,5 @@
 package com.example.labs.service;
 
-import com.example.labs.dto.AuthorDto;
 import com.example.labs.dto.AuthorFormDto;
 import com.example.labs.model.Author;
 import com.example.labs.repository.AuthorRepo;
@@ -13,18 +12,31 @@ import java.util.List;
 public class AuthorService {
     @Autowired
     AuthorRepo authorRepo;
-    @Autowired
-    BookRepo bookRepo;
 
     public List<AuthorFormDto> findAll() {
         return authorRepo.findAll().stream()
                 .map(a -> new AuthorFormDto(a.getId(), a.getFirstName(), a.getLastName(), a.getBirthYear(), a.getCountry()))
                 .toList();
     }
-    public void create(AuthorFormDto dto){
-        Author author = new Author(dto.getFirstName(), dto.getLastName(), dto.getBirthYear(), dto.getCountry());
-        authorRepo.save(author);
+    public AuthorFormDto create(AuthorFormDto dto){
+        Author author = new Author(
+                dto.getFirstName(),
+                dto.getLastName(),
+                dto.getBirthYear(),
+                dto.getCountry()
+        );
+
+        Author saved = authorRepo.save(author);
+
+        return new AuthorFormDto(
+                saved.getId(),
+                saved.getFirstName(),
+                saved.getLastName(),
+                saved.getBirthYear(),
+                saved.getCountry()
+        );
     }
+
 
     public AuthorFormDto findByIdForEdit(Long id) {
         Author author = authorRepo.findById(id)
@@ -39,7 +51,7 @@ public class AuthorService {
         );
     }
 
-    public void update(Long id, AuthorFormDto dto) {
+    public AuthorFormDto update(Long id, AuthorFormDto dto) {
         Author author = authorRepo.findById(id)
                 .orElseThrow(() -> new RuntimeException("Author not found"));
 
@@ -48,10 +60,21 @@ public class AuthorService {
         author.setBirthYear(dto.getBirthYear());
         author.setCountry(dto.getCountry());
 
-        authorRepo.save(author);
+        Author saved = authorRepo.save(author);
+
+        return new AuthorFormDto(
+                saved.getId(),
+                saved.getFirstName(),
+                saved.getLastName(),
+                saved.getBirthYear(),
+                saved.getCountry()
+        );
     }
 
     public void deleteById(Long id) {
+        if (!authorRepo.existsById(id)) {
+            throw new RuntimeException("Book not found");
+        }
         authorRepo.deleteById(id);
     }
 }
